+++ lib/features/digital_guide/tabs/structure/presentation/views/region_data_sliver_list_item.dart
@@ -1,0 +1,11 @@
+class RegionDataSliverListItem {
+  final String Function(int) text;
+  final Future<void> Function(int index)? onTap;
+  final int itemCount;
+
+  RegionDataSliverListItem({
+    required this.text,
+    this.onTap,
+    required this.itemCount,
+  });
+}
